+++ tests/app/Project/Console/ApplicationTestCase.php
@@ -4,6 +4,7 @@ namespace Project\Console;
 use Opulence\Applications\Application;
 use Opulence\Applications\Bootstrappers\ApplicationBinder;
 use Opulence\Framework\Testing\PhpUnit\Console\ApplicationTestCase as BaseTestCase;
+use Opulence\Ioc\IContainer;
 
 /**
  * Defines the console application test case
@@ -21,11 +22,13 @@ class ApplicationTestCase extends BaseTestCase
     /**
      * @inheritdoc
      */
-    protected function setApplication()
+    protected function setApplicationAndIocContainer()
     {
         /** @var Application $application */
         require __DIR__ . "/../../../../bootstrap/start.php";
         $this->application = $application;
+        /** @var IContainer $container */
+        $this->container = $container;
 
         /**
          * ----------------------------------------------------------